central_server/requirements.txt
fastapi==0.115.7
pycryptodome==3.21.0
pydantic==2.10.6
pydantic[email]
Requests==2.32.3
SQLAlchemy==2.0.37
uvicorn
websockets
pyjwt
bcrypt